Make mutation sorting more clear
Summary:
This is more insurance against future changes than fixing any existing bugs. In a very small number of cases it looks like this sorting isn't working, for reasons that are not easily reproducible (way less than 1 in 10000 times for me, if that) and not obvious, since we're sorting vectors in a canonical and straightforward way.

Again, this insures that logic won't change in the future.

Changelog: [Internal]

Reviewed By: mdvacca

Differential Revision: D27492936

fbshipit-source-id: 7f47e44ac1101915e1a0433c8f0a50a3c6a0c7b3